% Auto-generated: do not edit by hand
\name{''PivotTable}

\alias{''PivotTable}

\title{PivotTable component}

\description{
PivotTable is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''PivotTable(id=NULL, cols=NULL, data=NULL, label=NULL, rows=NULL,
value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{cols}{Logical | numeric | character | named list | unnamed list. The cols displayed in the input.}

\item{data}{Logical | numeric | character | named list | unnamed list. The data displayed in the input.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{rows}{Logical | numeric | character | named list | unnamed list. The row displayed in the input.}

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

